added serializer, changed sum integer to float

+++ db/migrate/20181128193945_create_spending_table.rb
@@ -2,12 +2,12 @@ class CreateSpendingTable < ActiveRecord::Migration[5.2]
   def change
     create_table :spendings do |t|
       t.integer :user_id
-      t.integer :rent
-      t.integer :food
-      t.integer :utilities
-      t.integer :clothes
-      t.integer :travel
-      t.integer :other
+      t.float :rent
+      t.float :food
+      t.float :utilities
+      t.float :clothes
+      t.float :travel
+      t.float :other
     end
   end
 end
